Set the initial values to 0 instead of 1.
Even though the first step is 1, if we initialize the values to 1
the current step won't be updated when we receive the first step
from the backend, and the graphs will start plotting from step 2.
Setting both the current and the max to 0, might also cause the
timer to pause, so add a check that the max is greated than 1.